chore(SFS): rename sfs resources and data sources to sfsturbo
