+++ huaweicloud/services/sfsturbo/data_source_huaweicloud_sfs_turbo_data_tasks.go
@@ -1,5 +1,5 @@
 // Generated by PMS #250
-package sfs
+package sfsturbo
 
 import (
 	"context"
