+++ huaweicloud/services/sfsturbo/data_source_huaweicloud_sfs_turbo_du_tasks.go
@@ -1,5 +1,5 @@
 // Generated by PMS #237
-package sfs
+package sfsturbo
 
 import (
 	"context"
