+++ huaweicloud/services/sfsturbo/data_source_huaweicloud_sfs_turbo_obs_targets.go
@@ -1,5 +1,5 @@
 // Generated by PMS #82
-package sfs
+package sfsturbo
 
 import (
 	"context"
