+++ huaweicloud/services/sfsturbo/data_source_huaweicloud_sfs_turbo_perm_rules.go
@@ -1,5 +1,5 @@
 // Generated by PMS #248
-package sfs
+package sfsturbo
 
 import (
 	"context"
