+++ huaweicloud/services/sfsturbo/data_source_huaweicloud_sfs_turbos.go
@@ -1,4 +1,4 @@
-package sfs
+package sfsturbo
 
 import (
 	"context"
