+++ huaweicloud/services/sfsturbo/resource_huaweicloud_sfs_turbo_obs_target.go
@@ -1,4 +1,4 @@
-package sfs
+package sfsturbo
 
 import (
 	"context"
